+++ inst/doc/tmb-gotchas.R
@@ -1,4 +1,0 @@
-## ---- message=FALSE, echo=FALSE-----------------------------------------------
-library(gadget3)
-library(magrittr)
-