+++ tests/test-suitability.R
@@ -23,6 +23,67 @@ suit_report_action <- function (pred_stock, stock, suit_f, run_at = 99) {
     return(as.list(out))
 }
 
+ok_group("g3_suitability_andersen", {
+    nondiff_andersen <- function (p0, p1, p2, p3 = p4, p4, p5, stock__midlen) {
+        p0 + p2 * exp(-(log(p5/stock__midlen) - p1)^2 / ifelse(log(p5/stock__midlen) <= p1, p4, p3))
+    }
+    g3_andersen <- function (p0, p1, p2, p3 = p4, p4, p5, stock__midlen) {
+        g3_eval(g3_suitability_andersen(p0, p1, p2, p3, p4, p5), stock__midlen = stock__midlen)
+    }
+
+    params <- list(p0 = 0, p1 = 0.659, p2 = 1, p3 = 0.15, p4 = 9942, p5 = 120, stock__midlen = 1:120)
+    ok(ut_cmp_equal(
+        do.call(g3_andersen, params),
+        do.call(nondiff_andersen, params),
+        tolerance = 1e-6), paste0("g3_suitability_andersen matches ", deparse1(params)))
+})
+
+ok_group("g3_suitability_andersenfleet", {
+    nondiff_andersenfleet <- function (
+            lg,
+            param.fish.andersen.p0 = 0,
+            param.fish.andersen.p1 = log(2),
+            param.fish.andersen.p2 = 1,
+            param.fish.andersen.p3_exp = 0.1,
+            param.fish.andersen.p4_exp = 0.1) {
+        p0 <- param.fish.andersen.p0
+        p1 <- param.fish.andersen.p1
+        p2 <- param.fish.andersen.p2
+        p3 <- exp(param.fish.andersen.p3_exp)
+        p4 <- exp(param.fish.andersen.p4_exp)
+        # Work out open-ended midlen, use maximum for p5
+        dl <- diff(lg) / 2 ; dl <- c(dl, dl[[length(dl)]])
+        stock__midlen <- lg + dl
+        p5 <- max(stock__midlen)
+
+        p0 + p2 * ifelse(
+            log(p5/stock__midlen) <= p1,
+            exp(-(log(p5/stock__midlen) - p1)**2/p4),
+            exp(-(log(p5/stock__midlen) - p1)**2/p3))
+    }
+    g3_andersenfleet <- function (lg, ...) {
+        g3_eval(
+            g3_suitability_andersenfleet(),
+            stock=g3_stock('fish', lg),
+            ...)
+    }
+
+    params <- list(lg = seq(100, 500, by = 100))
+    ok(ut_cmp_equal(
+        as.numeric(do.call(g3_andersenfleet, params)),
+        do.call(nondiff_andersenfleet, params),
+        tolerance = 1e-6), paste0("g3_suitability_andersen matches ", deparse1(params)))
+    params <- list(lg = seq(100, 500, by = 100), param.fish.andersen.p3_exp = -Inf)
+    ok(ut_cmp_equal(
+        as.numeric(do.call(g3_andersenfleet, params)),
+        do.call(nondiff_andersenfleet, params),
+        tolerance = 1e-6), paste0("g3_suitability_andersen matches ", deparse1(params)))
+    params <- list(lg = seq(100, 500, by = 100), param.fish.andersen.p4_exp = 0.999)
+    ok(ut_cmp_equal(
+        as.numeric(do.call(g3_andersenfleet, params)),
+        do.call(nondiff_andersenfleet, params),
+        tolerance = 1e-6), paste0("g3_suitability_andersen matches ", deparse1(params)))
+})
 
 fleet_stock <- g3_fleet('fleet') %>% g3s_livesonareas(1:3)
 pred_stock <- g3_stock('pred', c(10, 20, 30, 40, 50, 75, 100)) %>%
@@ -55,6 +116,12 @@ actions <- list(
         p1 = ~g3_param("p_andersen_p1"),
         p2 = ~g3_param("p_andersen_p2"),
         p4 = ~g3_param("p_andersen_p4"))),
+    suit_report_action(pred_stock, input_stock, g3_suitability_andersen(
+        p0 = ~g3_param("p_andersen_ne_p0"),
+        p1 = ~g3_param("p_andersen_ne_p1"),
+        p2 = ~g3_param("p_andersen_ne_p2"),
+        p3 = ~g3_param("p_andersen_ne_p3"),
+        p4 = ~g3_param("p_andersen_ne_p4"))),
     suit_report_action(fleet_stock, input_stock, g3_suitability_gamma(
         ~g3_param("p_gamma_alpha"),
         ~g3_param("p_gamma_beta"),